WIP for grid-based nav
